move files from /lib dockerignored folder
