+++ apps/web/src/tchap/components/views/dialogs/TchapCreateRoomDialog.tsx
@@ -26,7 +26,7 @@ import BaseDialog from "~tchap-web/src/components/views/dialogs/BaseDialog";
 import TchapUtils from "../../../util/TchapUtils";
 import TchapRoomTypeSelector from "../elements/TchapRoomTypeSelector";
 import { TchapRoomType } from "../../../@types/tchap";
-import TchapCreateRoom from "../../../lib/createTchapRoom";
+import TchapCreateRoom from "../../../ext/createTchapRoom";
 
 // We leave the same props as Element's version, to avoid unknown props warnings.
 interface IProps {
