+++ apps/web/test/unit-tests/tchap/createRoom-test.ts
@@ -8,11 +8,11 @@ import {
     stubClient,
 } from "~tchap-web/test/test-utils";
 import createRoom from "~tchap-web/src/createRoom";
-import ExternalAccountHandler from "~tchap-web/src/tchap/lib/ExternalAccountHandler";
+import ExternalAccountHandler from "~tchap-web/src/tchap/ext/ExternalAccountHandler";
 import SdkConfig, { type ConfigOptions } from "~tchap-web/src/SdkConfig";
 import Modal from "~tchap-web/src/Modal";
 
-jest.mock("~tchap-web/src/tchap/lib/ExternalAccountHandler.ts");
+jest.mock("~tchap-web/src/tchap/ext/ExternalAccountHandler.ts");
 
 describe("createRoom", () => {
     mockPlatformPeg();
