+++ apps/web/test/unit-tests/tchap/lib/createTchapRoom-test.ts
@@ -1,5 +1,5 @@
 import { TchapRoomType } from "../../../../src/tchap/@types/tchap";
-import TchapCreateRoom from "../../../../src/tchap/lib/createTchapRoom";
+import TchapCreateRoom from "../../../../src/tchap/ext/createTchapRoom";
 
 // todo(estellecomment) : these tests are redundant with TchapCreateRoomDialog-test. Either remove them or modify TchapCreateRoomDialog
 // to not edit state directly.
